# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de libros",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/books": {
            "get": {
                "summary": "Mostrar todos los libros",
                "operationId": "a9d13283dfab262e97af3c914d51b7a8",
                "responses": {
                    "200": {
                        "description": "Operación exitosa: Muestra todos los libros existentes."
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "summary": "Crear un nuevo libro",
                "operationId": "47047e95a700b3271b697e57d0965191",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Titulo de libro",
                                        "type": "string",
                                        "example": "El señor de los anillos"
                                    },
                                    "author": {
                                        "description": "Autor del libro",
                                        "type": "string",
                                        "example": "J.R.R. Tolkien"
                                    },
                                    "year": {
                                        "description": "Año de publicación",
                                        "type": "string",
                                        "example": "1954"
                                    },
                                    "genre": {
                                        "description": "Género del libro",
                                        "type": "string",
                                        "example": "Fantasía"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Libro creado correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/books/{id}": {
            "get": {
                "summary": "Obtener un libro por ID",
                "description": "Retorna un libro específico basado en su ID",
                "operationId": "2fdd668eb661591758aff4c56ca9b66a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Libro recuperado correctamente"
                    },
                    "404": {
                        "description": "Libro no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "put": {
                "summary": "Actualizar un libro por ID",
                "description": "Actualiza un libro especifico",
                "operationId": "f5ca658c501a803d68655af0365e3802",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Libro actualizado correctamente"
                    },
                    "404": {
                        "description": "Libro no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "summary": "Elimianr un libro por ID",
                "description": "Elimina un libro en especifico",
                "operationId": "1a9c26667c32b8ed5e5770d7a1cc9d5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Libro eliminado correctamente"
                    },
                    "404": {
                        "description": "Libro no encontrado"
                    },
                    "505": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        }
    }
}